requirements.txt
discord.py
playwright==1.49.1
python-dotenv